feat: additional debugging logs
